+++ packages/rsocket-router/src/router/ReactiveSocketRouter.ts
@@ -3,7 +3,7 @@
  * to expose reactive websocket stream in an interface similar to
  * other Journey micro routers.
  */
-import { errors, logger } from '@powersync/lib-services-framework';
+import { ErrorCode, errors, logger } from '@powersync/lib-services-framework';
 import * as http from 'http';
 import { Payload, RSocketServer } from 'rsocket-core';
 import * as ws from 'ws';
@@ -17,6 +17,14 @@ import {
   ReactiveSocketRouterOptions,
   SocketResponder
 } from './types.js';
+
+export interface ReactiveStreamRequest {
+  payload: Payload;
+  metadataMimeType: string;
+  dataMimeType: string;
+  initialN: number;
+  responder: SocketResponder;
+}
 
 export class ReactiveSocketRouter<C> {
   constructor(protected options?: ReactiveSocketRouterOptions<C>) {}
@@ -72,10 +80,16 @@ export class ReactiveSocketRouter<C> {
           // Throwing an exception in this context will be returned to the client side request
           if (!payload.metadata) {
             // Meta data is required for endpoint handler path matching
-            throw new errors.AuthorizationError('No context meta data provided');
+            throw new errors.AuthorizationError(ErrorCode.PSYNC_S2101, 'No context meta data provided');
           }
 
-          const context = await params.contextProvider(payload.metadata!);
+          const metadataMimeType = payload.metadataMimeType;
+          const dataMimeType = payload.dataMimeType;
+
+          const context = await params.contextProvider({
+            mimeType: payload.metadataMimeType,
+            contents: payload.metadata!
+          });
 
           return {
             // RequestStream is currently the only supported connection type
@@ -84,13 +98,17 @@ export class ReactiveSocketRouter<C> {
               const abortController = new AbortController();
 
               // TODO: Consider limiting the number of active streams per connection to prevent abuse
-              handleReactiveStream(context, { payload, initialN, responder }, observer, abortController, params).catch(
-                (ex) => {
-                  logger.error(ex);
-                  responder.onError(ex);
-                  responder.onComplete();
-                }
-              );
+              handleReactiveStream(
+                context,
+                { payload, initialN, responder, dataMimeType, metadataMimeType },
+                observer,
+                abortController,
+                params
+              ).catch((ex) => {
+                logger.error(ex);
+                responder.onError(ex);
+                responder.onComplete();
+              });
               return {
                 cancel: () => {
                   abortController.abort();
@@ -115,11 +133,7 @@ export class ReactiveSocketRouter<C> {
 
 export async function handleReactiveStream<Context>(
   context: Context,
-  request: {
-    payload: Payload;
-    initialN: number;
-    responder: SocketResponder;
-  },
+  request: ReactiveStreamRequest,
   observer: SocketRouterObserver,
   abortController: AbortController,
   params: CommonParams<Context>
@@ -137,7 +151,10 @@ export async function handleReactiveStream<Context>(
     return exitWithError(new errors.ValidationError('Metadata is not provided'));
   }
 
-  const meta = await params.metaDecoder(metadata);
+  const meta = await params.metaDecoder({
+    mimeType: request.metadataMimeType,
+    contents: metadata
+  });
 
   const { path } = meta;
 
@@ -148,7 +165,14 @@ export async function handleReactiveStream<Context>(
   }
 
   const { handler, authorize, validator, decoder = params.payloadDecoder } = route;
-  const requestPayload = await decoder(payload.data || undefined);
+  const requestPayload = await decoder(
+    payload.data
+      ? {
+          contents: payload.data,
+          mimeType: request.dataMimeType
+        }
+      : undefined
+  );
 
   if (validator) {
     const isValid = validator.validate(requestPayload);
@@ -166,7 +190,9 @@ export async function handleReactiveStream<Context>(
       responder
     });
     if (!isAuthorized.authorized) {
-      return exitWithError(new errors.AuthorizationError(isAuthorized.errors));
+      return exitWithError(
+        isAuthorized.error ?? new errors.AuthorizationError(ErrorCode.PSYNC_S2101, 'Authorization failed')
+      );
     }
   }