+++ packages/service-core/src/routes/endpoints/route-endpoints-index.ts
@@ -1,5 +1,6 @@
 export * from './admin.js';
 export * from './checkpointing.js';
+export * from './probes.js';
 export * from './socket-route.js';
 export * from './sync-rules.js';
 export * from './sync-stream.js';