+++ packages/service-core/src/routes/router.ts
@@ -1,4 +1,4 @@
-import { router } from '@powersync/lib-services-framework';
+import { router, ServiceError } from '@powersync/lib-services-framework';
 import type { JwtPayload } from '../auth/auth-index.js';
 import { ServiceContext } from '../system/ServiceContext.js';
 import { RouterEngine } from './RouterEngine.js';
@@ -6,7 +6,8 @@ import { RouterEngine } from './RouterEngine.js';
 /**
  * The {@link RouterEngine} must be provided for these routes
  */
-export type RouterServiceContext = ServiceContext & { routerEngine: RouterEngine };
+export type RouterServiceContext = ServiceContext;
+
 /**
  * Common context for routes
  */
@@ -16,7 +17,7 @@ export type Context = {
   service_context: RouterServiceContext;
 
   token_payload?: JwtPayload;
-  token_errors?: string[];
+  token_error?: ServiceError;
   /**
    * Only on websocket endpoints.
    */