+++ packages/service-errors/src/errors.ts
@@ -159,16 +159,44 @@ export class ReplicationAbortedError extends ServiceError {
   }
 }
 
-export class AuthorizationError extends ServiceError {
-  static readonly CODE = ErrorCode.PSYNC_S2101;
+export class UnsupportedMediaType extends ServiceError {
+  static readonly CODE = ErrorCode.PSYNC_S2004;
 
   constructor(errors: any) {
     super({
-      code: AuthorizationError.CODE,
-      status: 401,
-      description: 'Authorization failed',
+      code: UnsupportedMediaType.CODE,
+      status: 415,
+      description: 'Unsupported Media Type',
       details: errors
     });
+  }
+}
+
+export class AuthorizationError extends ServiceError {
+  /**
+   * String describing the token. Does not contain the full token, but may help with debugging.
+   * Safe for logs.
+   */
+  tokenDetails: string | undefined;
+  /**
+   * String describing related configuration. Should never be returned to the client.
+   * Safe for logs.
+   */
+  configurationDetails: string | undefined;
+
+  constructor(
+    code: ErrorCode,
+    description: string,
+    options?: { tokenDetails?: string; configurationDetails?: string; cause?: any }
+  ) {
+    super({
+      code,
+      status: 401,
+      description
+    });
+    this.cause = options?.cause;
+    this.tokenDetails = options?.tokenDetails;
+    this.configurationDetails = options?.configurationDetails;
   }
 }