+++ packages/service-core-tests/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @powersync/service-core-tests
 
+## 0.9.3
+
+### Patch Changes
+
+- 23ec406: Fix has_more and other data batch metadata
+- Updated dependencies [08f6ae8]
+- Updated dependencies [23ec406]
+- Updated dependencies [64e51d1]
+  - @powersync/service-core@1.11.3
+
 ## 0.9.2
 
 ### Patch Changes